Provide a simpler html start file for trying your own data
There's a lot to take in with DeepScatter, especially in the `index.html` file. Web dev is my day job, & it took me a good while to weed through it all & figure out what I needed for my own data. This file provides a good place to just see the basics.